move fromReadableSize to parseFileSize
